+++ Flow/src/utopia/flow/view/immutable/eventful/ChangeFuture.scala
@@ -1,14 +1,65 @@
 package utopia.flow.view.immutable.eventful
 
 import utopia.flow.async.AsyncExtensions._
+import utopia.flow.collection.CollectionExtensions._
 import utopia.flow.event.listener.{ChangeDependency, ChangeListener}
 import utopia.flow.event.model.ChangeEvent
+import utopia.flow.util.logging.Logger
+import utopia.flow.view.mutable.async.VolatileOption
 import utopia.flow.view.template.eventful.Changing
 
 import scala.concurrent.{ExecutionContext, Future}
+import scala.util.{Failure, Success, Try}
 
 object ChangeFuture
 {
+	// TODO: There should be some easy way to access the optimized constructors without repeating a lot of code
+	
+	/**
+	  * Creates a new future with change events
+	  * @param placeHolder A placeholder value stored until this future completes
+	  * @param future A future
+	  * @param merge A merge function that accepts the placeholder value and the future result (as a Try),
+	  *              yielding a new value to store.
+	  *              Will be called on both successful and failed resolves.
+	  * @param exc Implicit execution context.
+	  * @tparam A Type of stored value
+	  * @tparam F Type of future result, when successful
+	  * @return A new change future
+	  */
+	def merging[A, F](placeHolder: A, future: Future[F])(merge: (A, Try[F]) => A)(implicit exc: ExecutionContext) =
+		new ChangeFuture[A, F](placeHolder, future)(merge)
+	
+	/**
+	  * Creates a new change future
+	  * @param placeHolder A placeholder value returned until the future resolves (successfully)
+	  * @param future A future
+	  * @param exc Implicit execution context
+	  * @tparam A Type of held value
+	  * @return A new change future.
+	  *         This future will continue to contain the placeholder value if the future fails to resolve.
+	  */
+	def apply[A](placeHolder: A, future: Future[A])(implicit exc: ExecutionContext) =
+		merging[A, A](placeHolder, future) { (p, result) => result.getOrElse(p) }
+	
+	/**
+	  * Creates a new change future. Failed future resolves are logged.
+	  * @param placeHolder A placeholder value returned until the future resolves (successfully)
+	  * @param future      A future
+	  * @param exc         Implicit execution context
+	  * @param log A logging implementation for recording cases where the future fails
+	  * @tparam A Type of held value
+	  * @return A new change future.
+	  *         This future will continue to contain the placeholder value if the future fails to resolve.
+	  */
+	def caching[A](placeHolder: A, future: Future[A])(implicit exc: ExecutionContext, log: Logger) =
+		merging[A, A](placeHolder, future) { (p, result) =>
+			result.getOrMap { error =>
+				log(error)
+				p
+			}
+		}
+	
 	/**
 	  * Wraps a future into a change future
 	  * @param future A future to wrap
@@ -20,7 +71,7 @@ object ChangeFuture
 	def wrap[A](future: Future[A], placeHolder: => A)(implicit exc: ExecutionContext) =
 		future.current match {
 			case Some(v) => Fixed(v)
-			case None => new ChangeFuture[A](placeHolder, future)
+			case None => apply[A](placeHolder, future)
 		}
 	
 	/**
@@ -32,6 +83,26 @@ object ChangeFuture
 	  */
 	def noneUntilCompleted[A](future: Future[A])(implicit exc: ExecutionContext) =
 		wrap(future.map { Some(_) }, None)
+	
+	/**
+	  * Creates a change future that will contain Left until successfully resolved and Right once (if) so.
+	  * Logs possible errors.
+	  * @param placeHolder A placeholder value to return until the future successfully resolves.
+	  * @param future A future
+	  * @param exc Implicit execution context
+	  * @param log A logging implementation
+	  * @tparam A Type of future result & placeholder type
+	  * @return A new change future
+	  */
+	def eitherCaching[A](placeHolder: A, future: Future[A])(implicit exc: ExecutionContext, log: Logger) =
+		merging[Either[A, A], A](Left(placeHolder), future) { (p, result) =>
+			result match {
+				case Success(v) => Right(v)
+				case Failure(error) =>
+					log(error)
+					p
+			}
+		}
 }
 
 /**
@@ -39,36 +110,40 @@ object ChangeFuture
   * @author Mikko Hilpinen
   * @since 9.12.2020, v1.9
   */
-class ChangeFuture[A](placeHolder: A, val future: Future[A])(implicit exc: ExecutionContext) extends Changing[A]
+class ChangeFuture[A, F](placeHolder: A, val future: Future[F])(mergeResult: (A, Try[F]) => A)
+                        (implicit exc: ExecutionContext)
+	extends Changing[A]
 {
 	// ATTRIBUTES	------------------------------
 	
 	private var dependencies = Vector[ChangeDependency[A]]()
 	private var listeners = Vector[ChangeListener[A]]()
-	private var _value = placeHolder
+	
+	private val resultPointer = VolatileOption[A]()
 	
 	
 	// INITIAL CODE	------------------------------
 	
 	// Changes value when future completes
-	future.foreach { v =>
-		if (v != _value)
-		{
-			val event = ChangeEvent(_value, v)
-			// Updates local value
-			_value = v
-			
+	future.onComplete { result =>
+		val v = mergeResult(placeHolder, result)
+		// Updates local value
+		resultPointer.setOne(v)
+		
+		// Generates change events, if needed
+		if (v != placeHolder) {
+			val event = ChangeEvent(placeHolder, v)
 			// Informs the dependencies first
 			val afterEffects = dependencies.flatMap { _.beforeChangeEvent(event) }
 			// Then the listeners
 			listeners.foreach { _.onChangeEvent(event) }
 			// Finally performs dependency after-effects
-			afterEffects.foreach { _() }
-			
-			// Forgets about the listeners and dependencies afterwards
-			listeners = Vector()
-			dependencies = Vector()
+			afterEffects.foreach { _ () }
 		}
+		
+		// Forgets about the listeners and dependencies afterwards
+		listeners = Vector()
+		dependencies = Vector()
 	}
 	
 	
@@ -81,6 +156,8 @@ class ChangeFuture[A](placeHolder: A, val future: Future[A])(implicit exc: Execu
 	
 	
 	// IMPLEMENTED	------------------------------
+	
+	override def value = resultPointer.value.getOrElse(placeHolder)
 	
 	override def isChanging = !isCompleted
 	
@@ -99,36 +176,25 @@ class ChangeFuture[A](placeHolder: A, val future: Future[A])(implicit exc: Execu
 	override def addDependency(dependency: => ChangeDependency[A]) = if (isChanging) dependencies :+= dependency
 	override def removeDependency(dependency: Any) = dependencies = dependencies.filterNot { _ == dependency }
 	
-	override def futureWhere(valueCondition: A => Boolean) =
-		future.flatMap { v => if (valueCondition(v)) Future.successful(v) else Future.never }
-	override def nextFutureWhere(valueCondition: A => Boolean) =
-		if (isCompleted) Future.never else futureWhere(valueCondition)
-	
-	override def findMapFuture[B](f: A => Option[B]) =
-		future.flatMap { v => f(v) match {
-			case Some(v2) => Future.successful(v2)
-			case None => Future.never
-		} }
 	override def findMapNextFuture[B](f: A => Option[B]) =
 		if (isCompleted) Future.never else findMapFuture(f)
 	
-	override def map[B](f: A => B) =
-		if (isCompleted) Fixed(f(value)) else new ChangeFuture[B](f(placeHolder), future.map(f))
-	
+	override def map[B](f: A => B) = {
+		if (isCompleted)
+			Fixed(f(value))
+		else
+			new ChangeFuture[B, F](f(placeHolder), future)((_, result) => f(mergeResult(value, result)))
+	}
 	override def mergeWith[B, R](other: Changing[B])(f: (A, B) => R) = {
 		if (other.isChanging)
 			MergeMirror.of(this, other)(f)
 		else
 			map { f(_, other.value) }
 	}
-	
-	override def lazyMergeWith[B, R](other: Changing[B])(f: (A, B) => R) =
-	{
+	override def lazyMergeWith[B, R](other: Changing[B])(f: (A, B) => R) = {
 		if (other.isChanging)
 			LazyMergeMirror.of(this, other)(f)
 		else
 			lazyMap { f(_, other.value) }
 	}
-	
-	override def value = _value
 }